.github/workflows/requirements.txt
gitpython==3.1.24
openai==0.10.2